Add Zuul installation guide (#959)
Since the successful installation of a Zuul deployment is a bit more
complicated than simply applying the service role, we document it in a
dedicated chapter of the documentation.

Make zuul variables role specific:

Role variable names should have the role name as prefix in order to
avoid conflicts with other roles, amend the zuul role accordingly.