Add MQTT connection to the Zuul configuration
To view Zuul reports in the Matrix chat, we must use Zuul MQTT reports.
Since the direct path from MQTT to Matrix is unsupported,
we send MQTT reports through the zuul-mqtt-matrix-bridge,
which acts as a relay, transferring MQTT reports from Zuul to the Matrix chat.

This commit adds the MQTT connection to the Zuul configuration file.

Signed-off-by: Matej Feder <matej.feder@dnation.cloud>